% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_leverage_lpi.R
\name{calc_lambdas}
\alias{calc_lambdas}
\title{calc_lambdas - calculate interannual differences from an index}
\usage{
calc_lambdas(lpi)
}
\arguments{
\item{lpi}{- The index to calculate lambdas from}
}
\value{
Returns lamdbas for the lpi given
}
\description{
calc_lambdas - calculate interannual differences from an index
}
