% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalCSDev.R
\name{CalcSDev}
\alias{CalcSDev}
\title{CalcSDev - Calculates the difference operator on vector x to give an approximate
second derivative curve for x.}
\usage{
CalcSDev(x, h, d, interval)
}
\arguments{
\item{x}{the vector to calculate differences on}

\item{h}{is the desired window size. Corresponds to the quantity r in eqns 9, 10, and 11 of the Ecology paper.  Usually h will be chosen to be the smallest possible (h=1) but try experimenting with h=2 and h=3 to see if it has much effect on the changepoints.}

\item{d}{d is the order of the difference operator: can be 2, 4, or 6, representing the second derivative estimates obtained from 2nd differences, 4th differences, or 6th differences (eqns 9, 10, and 11 in the Ecology paper)  d=6 should give the most accurate results, but again try experimenting with d=4 and d=2.}

\item{interval}{is the time elapsed between points of x, which will be treated as 1 unit: e.g. if x consists of records taken every 2 years, then 2 years is equal to 1 unit, and interval=2.  Usually interval=1.}
}
\value{
Returns the approx second derivative of a curve for x
}
\description{
CalcSDev - Calculates the difference operator on vector x to give an approximate
second derivative curve for x.
}
