% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_index.R
\name{calculate_index}
\alias{calculate_index}
\title{Calculate the index using the annual differences in DTempArray}
\usage{
calculate_index(DTempArray, fileindex, DSize, Group, Weightings,
  use_weightings, use_weightings_B, WeightingsB)
}
\arguments{
\item{DTempArray}{Array of DTemps (annual differences)}

\item{fileindex}{The index of the file that this index is for}

\item{DSize}{The size of the data in DTemp}

\item{Group}{Which group this file belongs to}

\item{Weightings}{What the weightings are for this group}

\item{use_weightings}{Whether or not to use weightings (level 1)}

\item{use_weightings_B}{Whether or not to use weightings (level 1)}

\item{WeightingsB}{What the weightingsB are for this group}
}
\value{
index - the calculated index
}
\description{
Calculate the index using the annual differences in DTempArray
}
