% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/debug_print.R
\name{debug_print}
\alias{debug_print}
\title{Function to print if the global variable 'DEBUG' is set}
\usage{
debug_print(text)
}
\arguments{
\item{text}{Text to print}
}
\description{
Function to print if the global variable 'DEBUG' is set
}

