% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_lambdas.R
\name{ggplot_lambdas}
\alias{ggplot_lambdas}
\title{Plots annual lambdas}
\usage{
ggplot_lambdas(lambda_values, col = "red", line_col = "blue",
  title = "", baseline = 1, baseline_color = "grey", ylims = c(0.9,
  1.1), ybreaks = 0.05, xlims = c(1970, 2014.5), trans = "identity",
  yrbreaks = 5)
}
\arguments{
\item{lambda_values}{- the dataframe containing lambda values}

\item{col}{- Color of plotted points, default="red"}

\item{line_col}{- Color of error bars, default="blue"}

\item{title}{- Title of plot}

\item{baseline}{- baseline to show on plot, default=1}

\item{baseline_color}{- color of baseline, default="grey"}

\item{ylims}{- y_limits, default=c(0.9, 1.1)}

\item{xlims}{- y_limits, default=c(1970, 2014.5)}

\item{trans}{- y-axis transformation, default="identity", could be "log"}

\item{yrbreaks}{- The spacing between x-axis tick marks}
}
\value{
Returns the calculated plot
}
\description{
Plots annual lambdas
}
