% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_lpi.R
\name{ggplot_lpi}
\alias{ggplot_lpi}
\title{Plots an index using ggplot (dataframe containing LPI_final, CI_low, CI_high)}
\usage{
ggplot_lpi(d, col = "darkblue", line_col = "white", title = "",
  ylims = c(0, 2), xlims = NULL, trans = "identity", yrbreaks = 5,
  lpi_breaks = 0.2)
}
\arguments{
\item{d}{- the dataframe containing colnames(LPI_final, CI_low, CI_high)}

\item{col}{- Color of plotted index ribbon (confidence intervals)}

\item{line_col}{- Color of mean line of plotting index}

\item{title}{- Title of plot}

\item{ylims}{- ylim of plot}

\item{xlims}{- xlim of plot}

\item{trans}{- The y-axis transformation, default is 'identitiy', but could be 'log'}

\item{yrbreaks}{- The spacing between x-axis tick marks}

\item{lpi_breaks}{- The spacing between y-axis tick marks}
}
\value{
Returns the calculated plot
}
\description{
Plots an index using ggplot (dataframe containing LPI_final, CI_low, CI_high)
}
