% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_lpi_withN.R
\name{ggplot_lpi_withN}
\alias{ggplot_lpi_withN}
\title{Plots an index with the number of points contributing using ggplot (dataframe containing LPI_final, CI_low, CI_high, LPI=number of points)}
\usage{
ggplot_lpi_withN(d, col = "black", title = "", y_min = 0,
  y_max = 2)
}
\arguments{
\item{d}{- the dataframe containing colnames(LPI_final, CI_low, CI_high, LPI)}

\item{col}{- Color of plotted index}

\item{title}{- Title of plot}

\item{y_min}{- y_min of plot}

\item{y_max}{- y_max of plot}
}
\value{
Returns the calculated plot
}
\description{
Plots an index with the number of points contributing using ggplot (dataframe containing LPI_final, CI_low, CI_high, LPI=number of points)
}
