% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_multi_lpi.R
\name{ggplot_lpis}
\alias{ggplot_lpis}
\title{Title}
\usage{
ggplot_lpis(lpis, names, title = "", col = "Set2", ylims = c(0.4, 2.5),
  trans = "identity")
}
\arguments{
\item{lpis}{- The list of lpis to plot}

\item{names}{- the names of the each lpi in the list}

\item{title}{- the title of the plots (same for each)}

\item{col}{- The RColorBrewer set to use. Default is "Set2"}

\item{ylims}{- The ylims of the plot}

\item{trans}{- The transformation to apply to the y-axis. Default is 'identity' could be 'log'}
}
\value{
Returns the calculated plot
}
\description{
Title
}

