% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_multi_lpi_2x2.R
\name{ggplot_multi_lpi_2x2}
\alias{ggplot_multi_lpi_2x2}
\title{ggplot_multi_lpi_2x2}
\usage{
ggplot_multi_lpi_2x2(lpis, names, title = "", col = "Blues")
}
\arguments{
\item{lpis}{- the list of lpis to plot}

\item{names}{- names for each lpi in the list}

\item{title}{- the title of the plot (same for each)}

\item{col}{- the RColorBrewer color group to use. Default is "Blues"}
}
\value{
Returns the calculated plot
}
\description{
ggplot_multi_lpi_2x2
}
