% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPI_v15_20150915.R
\name{LPIMain}
\alias{LPIMain}
\title{Main function for generating indices from an input file}
\usage{
LPIMain(infile, force_recalculation = DEFAULT_FORCE_RECALCULATION,
  use_weightings = DEFAULT_WEIGHTING,
  use_weightings_B = DEFAULT_WEIGHTING_B, title = "",
  CI_FLAG = DEFAULT_CI_FLAG, LEV_FLAG = DEFAULT_LEV_FLAG,
  SWITCH_PT_FLAG = DEFAULT_SWITCH_PT_FLAG,
  BOOT_STRAP_SIZE = DEFAULT_BOOT_STRAP_SIZE,
  save_plots = DEFAULT_SAVE_PLOTS, plot_lpi = DEFAULT_PLOT_LPI,
  goParallel = FALSE, MODEL_SELECTION_FLAG = DEFAULT_MODEL_SELECTION_FLAG,
  GAM_GLOBAL_FLAG = DEFAULT_GAM_GLOBAL_FLAG,
  DATA_LENGTH_MIN = DEFAULT_DATA_LENGTH_MIN,
  AVG_TIME_BETWEEN_PTS_MAX = DEFAULT_AVG_TIME_BETWEEN_PTS_MAX,
  GLOBAL_GAM_FLAG_SHORT_DATA_FLAG = DEFAULT_GLOBAL_GAM_FLAG_SHORT_DATA_FLAG,
  AUTO_DIAGNOSTIC_FLAG = DEFAULT_AUTO_DIAGNOSTIC_FLAG,
  LAMBDA_MIN = DEFAULT_LAMBDA_MIN, LAMBDA_MAX = DEFAULT_LAMBDA_MAX,
  ZERO_REPLACE_FLAG = DEFAULT_ZERO_REPLACE_FLAG,
  OFFSET_ALL = DEFAULT_OFFSET_ALL)
}
\arguments{
\item{infile}{Input file specifying the population files that should be included in the index}

\item{force_recalculation}{Whether the population annual differences should be recalucated (they are cached for a given file). Default=0}

\item{use_weightings}{Whether to use the first level of weightings ('Weightings') in the infile. Default=0}

\item{use_weightings_B}{Whether to use the first level of weightings ('WeightingsB') in the infile. Default=0}

\item{title}{The title. Default=""}

\item{CI_FLAG}{Flag to indicate whether or not to calculate confidence intervals (bootstrapping the index). Default=1}

\item{LEV_FLAG}{Flag to indicate wether or not to calculate species level leverage plots. Default=0}

\item{SWITCH_PT_FLAG}{Flag to indicate whether or not to calculate switching points. Default=0}

\item{BOOT_STRAP_SIZE}{If calculating CIs, how many bootstraps to use. Default=100}

\item{save_plots}{Should plots be saved. Default=1}

\item{plot_lpi}{Should plots be plotted. Default=1}

\item{goParallel}{Should the code be executing in parallel, Default=FALSE}

\item{MODEL_SELECTION_FLAG}{Default=0}

\item{GAM_GLOBAL_FLAG}{1 = process by GAM method, 0 = process by chain method. Default=1}

\item{DATA_LENGTH_MIN}{Minimum data length to include in calculations. Default=2}

\item{AVG_TIME_BETWEEN_PTS_MAX}{Maximum time between datapoint to include. Default=100}

\item{GLOBAL_GAM_FLAG_SHORT_DATA_FLAG}{# set this to 1 GAM model is also to be generated for the short time series else the log linear model will be used. Default=0}

\item{AUTO_DIAGNOSTIC_FLAG}{1=Automatically determine whether GAM models are good enough, 0=Manually ask for each. Default=1}

\item{LAMBDA_MIN}{Minimum lambda to include in calculations. Default=1}

\item{LAMBDA_MAX}{Minimum lambda to include in calculations. Default=-1}

\item{ZERO_REPLACE_FLAG}{0 = +minimum value; 1 = +1\% of mean value; 2 = +1. Default=2}

\item{OFFSET_ALL}{1 = Add offset to all values, to avoid log(0). Default=0}
}
\value{
lpi - A data frame containing an LPI and CIs if calculated
}
\description{
Before March 2009
--------
Version 9 R Amin Incorporated filtering of Lambda values outside a defined range
Incorporated adding 1) 1 \% mean of non zero values, 2) min value of non zero
values to all the values if there is a zero value in the time series Incorporated
time series of >= 2 are processed
--------
March 2009 - Ben has messed with the flags, so that more are up at the top (e.g.
Bootstrap size).  Functionality of the CI_FLAG improved, introduced a PLOT_MAX value
for the plotting, added commands for the zero replace flag.
--------
Version 12 (August 2010 - Stef changed some of the output file names: MethodFlag
-> PopProcessedChain, new_results -> 01_Results, so that file is always at the top
and therefore easy to find. Also changed code to output the Species lambda and
DTemp files as txt (NOTE: currently only works for one input text file). -1 in the
lambda file is now NA. Also, additional output file (Lambda.txt) created,
combining SpeciesName and SpeciesLambda and automatically inserting years
according to REF_YEAR.
--------
Version 13 - Stef has integrated minmax code at the end of
the LPI code
--------
Version 13 - update SD fixed lambda file error Version 13 update - SD
deleted ylim from graph and added automatic savePlot function
--------
Version 14 - July-October 2013, Robin Freeman
}
\details{
Reformatted code to improve readabilitiy
Added comments throughout
Reorganised some sections into new functions (ProcessFile, Bootstrap_lpi) to improve
readability and to enable paralisation
Split functions into separate files

Version 15 - Sept 2015 - Robin Freeman
 - Various updates, added associated plotting code (ggplot_lpi.R)
 - moved some parameters to the function call (# bootstraps)
 - Moved CalcLPI parameters into the function call and to the LPIMain function allowing them to be controlled
   from the function call

- Created lpi_options.R to store overall options
- Reorganised to now run with a parameter --- LPIMain(infile)

# Calculate an index using the population file specified in GlobalInFile.txt, calculating confidence intervals using 100 bootstraps
lpi_global <- LPIMain("GlobalInfile.txt", CI_FLAG=1, title="Global LPI", BOOT_STRAP_SIZE=100)
# Plot this global LPI
ggplot_lpi(lpi_global)
# Calculate an index using the population file specified in TropicalInfile.txt, calculating confidence intervals using 100 bootstraps
lpi_tropical <- LPIMain("TropicalInfile.txt", CI_FLAG=1, title="Tropical LPI", BOOT_STRAP_SIZE=100)
# Plot this tropical LPI
ggplot_lpi(lpi_tropical)
# Plot them together
ggplot_lpi_multi(list(lpi_global, lpi_tropical), names=c("global", "tropical"))
}

