% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lpi.R
\name{plot_lpi}
\alias{plot_lpi}
\title{plot_lpi}
\usage{
plot_lpi(index, REF_YEAR, PLOT_MAX, CI_FLAG = 0, lowerCI = 0,
  upperCI = 0, col = "black")
}
\arguments{
\item{index}{- The index to plot}

\item{REF_YEAR}{- The reference year of the plot (index == 1)}

\item{PLOT_MAX}{- The max y-value of the plot}

\item{CI_FLAG}{- whether confidence intervals are to be plotted}

\item{lowerCI}{- lower confidence interval values}

\item{upperCI}{- upper confidence interval values}

\item{col}{- The color of the plot. Default is "black"}
}
\description{
plot_lpi
}
