% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_lpi.R
\name{smooth_lpi}
\alias{smooth_lpi}
\title{smooth_lpi}
\usage{
smooth_lpi(d, N, smooth_type = "loess")
}
\arguments{
\item{d}{- the index to smooth}

\item{N}{- The size of the smoothing window}

\item{smooth_type}{- Type of smoothing. Default is 'loess', otherwise uses boxcar with ends pinned}
}
\value{
The smoothed LPI
}
\description{
smooth_lpi
}
