% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lpi.R
\name{summarise_lpi}
\alias{summarise_lpi}
\title{summarise_lpi - generates useful summaries for data within the given infile}
\usage{
summarise_lpi(infile)
}
\arguments{
\item{infile}{- The infile to summarise the data within}
}
\description{
summarise_lpi - generates useful summaries for data within the given infile
}
