% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_print.R
\name{debug_print}
\alias{debug_print}
\title{Function to print if the global variable 'DEBUG' is set}
\usage{
debug_print(verbose = FALSE, text)
}
\arguments{
\item{verbose}{Flag to control printing}

\item{text}{Text to print}
}
\description{
Function to print if the global variable 'DEBUG' is set
}
