% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LPIMain.R
\name{LPIMain}
\alias{LPIMain}
\title{Create an index from a number of datasets in a given input file}
\usage{
LPIMain(infile = "Infile.txt", REF_YEAR = 1970, PLOT_MAX = 2012,
  force_recalculation = 1, use_weightings = 0, use_weightings_B = 0,
  title = "", CI_FLAG = 1, LEV_FLAG = 0, SWITCH_PT_FLAG = 0,
  BOOT_STRAP_SIZE = 100, save_plots = 1, plot_lpi = 1,
  goParallel = FALSE, MODEL_SELECTION_FLAG = 0, GAM_GLOBAL_FLAG = 1,
  DATA_LENGTH_MIN = 2, AVG_TIME_BETWEEN_PTS_MAX = 100,
  GLOBAL_GAM_FLAG_SHORT_DATA_FLAG = 0, AUTO_DIAGNOSTIC_FLAG = 1,
  LAMBDA_MIN = -1, LAMBDA_MAX = 1, ZERO_REPLACE_FLAG = 2,
  OFFSET_ALL = 0, VERBOSE = TRUE)
}
\arguments{
\item{infile}{Input file specifying the population files that should be included in the index}

\item{REF_YEAR}{Reference year for index (when the index == 1). Default=1970}

\item{PLOT_MAX}{The final year of the index to plot. Default=2012,}

\item{force_recalculation}{Whether the population annual differences should be recalucated (they are cached for a given file). Default=0}

\item{use_weightings}{Whether to use the first level of weightings ('Weightings') in the infile. Default=0}

\item{use_weightings_B}{Whether to use the first level of weightings ('WeightingsB') in the infile. Default=0}

\item{title}{The title. Default=""}

\item{CI_FLAG}{Flag to indicate whether or not to calculate confidence intervals (bootstrapping the index). Default=1}

\item{LEV_FLAG}{Flag to indicate wether or not to calculate species level leverage plots. Default=0}

\item{SWITCH_PT_FLAG}{Flag to indicate whether or not to calculate switching points. Default=0}

\item{BOOT_STRAP_SIZE}{If calculating CIs, how many bootstraps to use. Default=100}

\item{save_plots}{Should plots be saved. Default=1}

\item{plot_lpi}{Should plots be plotted. Default=1}

\item{goParallel}{Should the code be executing in parallel, Default=FALSE}

\item{MODEL_SELECTION_FLAG}{Default=0}

\item{GAM_GLOBAL_FLAG}{1 = process by GAM method, 0 = process by chain method. Default=1}

\item{DATA_LENGTH_MIN}{Minimum data length to include in calculations. Default=2}

\item{AVG_TIME_BETWEEN_PTS_MAX}{Maximum time between datapoint to include. Default=100}

\item{GLOBAL_GAM_FLAG_SHORT_DATA_FLAG}{# set this to 1 GAM model is also to be generated for the short time series else the log linear model will be used. Default=0}

\item{AUTO_DIAGNOSTIC_FLAG}{1=Automatically determine whether GAM models are good enough, 0=Manually ask for each. Default=1}

\item{LAMBDA_MIN}{Minimum lambda to include in calculations. Default=1}

\item{LAMBDA_MAX}{Minimum lambda to include in calculations. Default=-1}

\item{ZERO_REPLACE_FLAG}{0 = +minimum value; 1 = +1\% of mean value; 2 = +1. Default=2}

\item{OFFSET_ALL}{1 = Add offset to all values, to avoid log(0). Default=0}

\item{VERBOSE}{Whether to print verbose information. Default=1}
}
\value{
lpi - A data frame containing an LPI and CIs if calculated
}
\description{
Create an index from a number of datasets in a given input file
}
\details{
Version 15 - Sept 2015 - Robin Freeman
 \itemize{
\item Various updates, added associated plotting code (ggplot_lpi.R)
 \item moved some parameters to the function call (# bootstraps)
 \item Moved CalcLPI parameters into the function call and to the LPIMain function allowing them to be controlled from the function call
 \item Created lpi_options.R to store overall options
 \item Reorganised to now run with main options as parameters --- LPIMain(infile, etc.)
 }

Version 14 - July-October 2013, Robin Freeman
\itemize{
\item Reformatted code to improve readabilitiy
\item Added comments throughout
\item Reorganised some sections into new functions (ProcessFile, Bootstrap_lpi) to improve readability and to enable paralisation
\item Split functions into separate files
}

Version 13 - update SD fixed lambda file error Version 13 update - SD
deleted ylim from graph and added automatic savePlot function

Version 13 - Stef has integrated minmax code at the end of
the LPI code

Version 12 (August 2010 - Stef changed some of the output file names: MethodFlag
-> PopProcessedChain, new_results -> 01_Results, so that file is always at the top
and therefore easy to find. Also changed code to output the Species lambda and
DTemp files as txt (NOTE: currently only works for one input text file). -1 in the
lambda file is now NA. Also, additional output file (Lambda.txt) created,
combining SpeciesName and SpeciesLambda and automatically inserting years
according to REF_YEAR.

March 2009 - Ben has messed with the flags, so that more are up at the top (e.g.
Bootstrap size).  Functionality of the CI_FLAG improved, introduced a PLOT_MAX value
for the plotting, added commands for the zero replace flag.

Before March 2009.... Version 9 R Amin Incorporated filtering of Lambda values outside a defined range
Incorporated adding 1) 1 \% mean of non zero values, 2) min value of non zero
values to all the values if there is a zero value in the time series Incorporated
time series of >= 2 are processed

# Calculate an index using the population file specified in GlobalInFile.txt, calculating confidence intervals using 100 bootstraps
lpi_global <- LPIMain("GlobalInfile.txt", CI_FLAG=1, title="Global LPI", BOOT_STRAP_SIZE=100)
# Plot this global LPI
ggplot_lpi(lpi_global)
# Calculate an index using the population file specified in TropicalInfile.txt, calculating confidence intervals using 100 bootstraps
lpi_tropical <- LPIMain("TropicalInfile.txt", CI_FLAG=1, title="Tropical LPI", BOOT_STRAP_SIZE=100)
# Plot this tropical LPI
ggplot_lpi(lpi_tropical)
# Plot them together
ggplot_lpi_multi(list(lpi_global, lpi_tropical), names=c("global", "tropical"))
}
\examples{
# Get example data from package
# Copy zipped data to local directory
file.copy(from=system.file("extdata", "example_data.zip", package = "rlpi"), to=".")
unzip("example_data.zip")

# Terrestrial LPI with equal weighting across classes and realms
# Default gives 100 boostraps (this will take a few minutes to run (on a 2014 Macbook))
terr_lpi <- LPIMain("terrestrial_class_realms_infile.txt")

# Nicer plot
ggplot_lpi(terr_lpi)

# Run same again, but used cached lambdas (force_recalculation == 0), and now weighted by class, but equal across realms (see infile for weights)
terr_lpi_b <- LPIMain("terrestrial_class_realms_infile.txt", force_recalculation=0, use_weightings=1)

# Putting the two LPIs together in a list
lpis <- list(terr_lpi, terr_lpi_b)
# And plotting them together should show identical means but with different CIs
ggplot_multi_lpi(lpis, xlims=c(1970, 2012))

# Can also plot these next to each other, and use some more meaningful titles
ggplot_multi_lpi(lpis, names=c("Weighted", "Unweighted"), xlims=c(1970, 2012), facet=TRUE)

# And can log the y-axis - need to set ylim as log(0) is -Inf
ggplot_multi_lpi(lpis, names=c("Weighted", "Unweighted"), xlims=c(1970, 2012), facet=TRUE, ylim=c(0.5, 2), trans="log")
}

